+++ docs/source/changelog.rst
@@ -35,9 +35,11 @@ Changelog
  - Support :class:`.Index` objects in :meth:`.DDLSession.create_table`.
 
  - Actually generate foreign keys upon table creation.
- 
+
  - Add :class:`.Serial`, :class:`.BigSerial`, and :class:`.SmallSerial` types to support automatic
      incrementation
+
+ - Add :class:`.UpsertQuery`.
 
 
 0.1.0 (released 2017-07-30)